# perf(fastrand): optimize Read